Extract abstract `Graph` class from `Workflow`
